# Conflicts:
#	src/main/java/de/uniks/beastopia/teaml/controller/ingame/EntityController.java
#	src/main/java/de/uniks/beastopia/teaml/controller/ingame/IngameController.java
#	src/main/java/de/uniks/beastopia/teaml/controller/ingame/TrainerController.java
#	src/main/java/de/uniks/beastopia/teaml/controller/ingame/encounter/StartFightNPCController.java
#	src/main/java/de/uniks/beastopia/teaml/service/DataCache.java
#	src/main/resources/de/uniks/beastopia/teaml/assets/lang_de.properties
#	src/main/resources/de/uniks/beastopia/teaml/assets/lang_en.properties
#	src/main/resources/de/uniks/beastopia/teaml/views/dark.css
#	src/test/java/de/uniks/beastopia/teaml/controller/ingame/EntityControllerTest.java
#	src/test/java/de/uniks/beastopia/teaml/controller/ingame/encounter/StartFightNPCControllerTest.java

+++ src/main/java/de/uniks/beastopia/teaml/rest/Chunk.java
@@ -3,7 +3,7 @@ package de.uniks.beastopia.teaml.rest;
 import java.util.List;
 
 public record Chunk(
-        List<Integer> data,
+        List<Long> data,
         int height,
         int width,
         int x,
